+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * :sparkles: Stringx.kt adds find, findLast, first, firstOrNull, last, lastOrNull, getOrElse, getOrNull, indexOfFirst, indexOfLast, substring, subSequence, iterator, iterable, indices, lastIndex method
 * :fire: CharIterator rename to CharRangeIterator, DoubleIterator rename to DoubleRangeIterator, FloatIterator rename to FloatRangeIterator,
     IntIterator rename to IntRangeIterator, LongIterator rename to LongRangeIterator, DateIterator rename to DateRangeIterator
+* :sparkles: Collectionx adds setOf
 
 
 ## v0.7.1